refine sidebar view, add selected state to line componment
