+++ src/style.css
@@ -71,22 +71,42 @@ td.new_line_number, td.old_line_number {
 	display: none
 }
 
+
+
+
 /* Styling for file view. TODO merge with app. */
 .file_line_number {
 	width: 40px;
 	border-right:1px solid #eaeaea;
 	color: #85878c;
-	padding: 1px;
+	/*padding: 1px;*/
+}
+
+.file_line_text{
+	border-style: solid;
+	border-width: 2px;
+	border-color: transparent;
+	padding: 0px;
 }
 
 .file_line_text:hover {
-	background-color: LightGrey ;
+	background-color: LightGrey;
 }
 
+.file_line_text.selected {
+	border-style: solid;
+  border-width: 2px;
+	border-color: LightGrey;
+	padding: 0px;
+}
+
+/* file view right-siderbar */
 .tests_viewer {
 	position: fixed;
 	top: 0;
+	bottom: 0;
 	right: 0;
-	width: 400px;
-	height: 100%;
+	width: 240px;
+	border-left: 2px solid #A0A0A0;
+	padding-left: 20px;
 }
